services/dispatcher/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
prometheus-client==0.21.0
pymongo==4.8.0
pyjwt==2.9.0

pytest==8.3.2
pytest-asyncio==0.24.0

